+++ shared/src/iosTest/kotlin/com/riffle/shared/reader/ReadiumSwiftNavigatorAnnotationTest.kt
@@ -42,7 +42,7 @@ class ReadiumSwiftNavigatorAnnotationTest {
         override fun snapshotLocatorJson(): String? = null
         override fun setLocatorCallback(callback: ((locatorJson: String) -> Unit)?) = Unit
         override fun setPageLoadCallback(callback: (() -> Unit)?) = Unit
-        override fun setTapCallback(callback: ((Float, Float, Float, Float) -> Unit)?) = Unit
+        override fun setTapCallback(callback: ((Double, Double, Double, Double) -> Unit)?) = Unit
         override fun setErrorCallback(callback: ((message: String) -> Unit)?) = Unit
         override fun setSelectionCallback(callback: ((selectionJson: String?) -> Unit)?) {
             selectionCallback = callback
